Create separate component for the Dialog and add unit test
